Use one module and avoid wrappers
